# openapi.json
{"openapi":"3.0.2","info":{"title":"FastAPI","version":"0.1.0"},"paths":{"/api/v1/vehicleDbc/dids/{vehicle}":{"get":{"tags":["Vehicle DBC"],"summary":"Get Vehicle Dids","operationId":"get_vehicle_dids_api_v1_vehicleDbc_dids__vehicle__get","parameters":[{"required":true,"schema":{"title":"Vehicle Id","type":"string"},"name":"vehicle_id","in":"query"}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/VehicleModelDIDs"}}}},"404":{"description":"Not found"},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/vehicleDbc/dids":{"post":{"tags":["Vehicle DBC"],"summary":"Add Update Vehicles Dids","operationId":"add_update_vehicles_dids_api_v1_vehicleDbc_dids_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/Body_add_update_vehicles_dids_api_v1_vehicleDbc_dids_post"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"title":"Response Add Update Vehicles Dids Api V1 Vehicledbc Dids Post","type":"array","items":{"$ref":"#/components/schemas/VehicleModelDIDs"}}}}},"404":{"description":"Not found"},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/dbc/upload-dbc":{"post":{"tags":["DBC CAN Message"],"summary":"Upload Dbc","operationId":"upload_dbc_api_v1_dbc_upload_dbc_post","parameters":[{"required":false,"schema":{"title":"Frame Id","type":"integer"},"name":"frame_id","in":"query"}],"requestBody":{"content":{"multipart/form-data":{"schema":{"$ref":"#/components/schemas/Body_upload_dbc_api_v1_dbc_upload_dbc_post"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"title":"Response Upload Dbc Api V1 Dbc Upload Dbc Post","anyOf":[{"type":"array","items":{"$ref":"#/components/schemas/CANFrame"}},{"type":"array","items":{"$ref":"#/components/schemas/FrameSignal"}}]}}}},"404":{"description":"Not found"},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/vehicleLogs/{vehcile_id}":{"get":{"tags":["Vehicle Logs"],"summary":"Get All Vehiclelogs","operationId":"get_all_vehicleLogs_api_v1_vehicleLogs__vehcile_id__get","parameters":[{"required":true,"schema":{"title":"Vehcile Id","type":"string"},"name":"vehcile_id","in":"path"},{"required":false,"schema":{"title":"Limit","type":"integer","default":100},"name":"limit","in":"query"},{"required":false,"schema":{"title":"Start Time","type":"string","format":"date-time"},"name":"start_time","in":"query"},{"required":false,"schema":{"title":"End Time","type":"string","format":"date-time"},"name":"end_time","in":"query"}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"title":"Response Get All Vehiclelogs Api V1 Vehiclelogs  Vehcile Id  Get","type":"array","items":{"$ref":"#/components/schemas/VehicleLogsObject"}}}}},"404":{"description":"Not found"},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/vehicleLogs/live-logs":{"post":{"tags":["Vehicle Logs"],"summary":"Get Live Logs","operationId":"get_live_logs_api_v1_vehicleLogs_live_logs_post","parameters":[{"required":false,"schema":{"title":"Timeout","type":"integer","default":100},"name":"timeout","in":"query"}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/Body_get_live_logs_api_v1_vehicleLogs_live_logs_post"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"title":"Response Get Live Logs Api V1 Vehiclelogs Live Logs Post","type":"array","items":{"$ref":"#/components/schemas/VehicleLogsObject"}}}}},"404":{"description":"Not found"},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/":{"get":{"summary":"Root","operationId":"root__get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}},"/health":{"get":{"summary":"Health Check","operationId":"health_check_health_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}}},"components":{"schemas":{"Body_add_update_vehicles_dids_api_v1_vehicleDbc_dids_post":{"title":"Body_add_update_vehicles_dids_api_v1_vehicleDbc_dids_post","required":["vehicle_ids","didsList"],"type":"object","properties":{"vehicle_ids":{"title":"Vehicle Ids","type":"array","items":{"type":"string"}},"didsList":{"title":"Didslist","type":"array","items":{"$ref":"#/components/schemas/VehicleDID"}}}},"Body_get_live_logs_api_v1_vehicleLogs_live_logs_post":{"title":"Body_get_live_logs_api_v1_vehicleLogs_live_logs_post","required":["vehicle_ids","didsList"],"type":"object","properties":{"vehicle_ids":{"title":"Vehicle Ids","type":"array","items":{"type":"string"}},"didsList":{"title":"Didslist","type":"array","items":{"$ref":"#/components/schemas/VehicleDID"}}}},"Body_upload_dbc_api_v1_dbc_upload_dbc_post":{"title":"Body_upload_dbc_api_v1_dbc_upload_dbc_post","required":["dbc_file"],"type":"object","properties":{"dbc_file":{"title":"Dbc File","type":"string","format":"binary"}}},"CANFrame":{"title":"CANFrame","required":["frame_id","name"],"type":"object","properties":{"frame_id":{"title":"Frame Id","type":"integer"},"name":{"title":"Name","type":"string"},"autosar":{"title":"Autosar","type":"string"},"protocol":{"title":"Protocol","type":"string"},"signals":{"title":"Signals","type":"array","items":{"$ref":"#/components/schemas/FrameSignal"}}}},"FrameSignal":{"title":"FrameSignal","required":["name","start","length","offset","type"],"type":"object","properties":{"name":{"title":"Name","type":"string"},"start":{"title":"Start","type":"integer"},"length":{"title":"Length","type":"integer"},"offset":{"title":"Offset","type":"integer"},"unit":{"title":"Unit","type":"string"},"type":{"title":"Type","type":"string"},"byte_order":{"title":"Byte Order","type":"string"},"min":{"title":"Min","type":"number"},"max":{"title":"Max","type":"number"},"receivers":{"title":"Receivers","type":"array","items":{"type":"string"}},"comment":{"title":"Comment","type":"string"}}},"HTTPValidationError":{"title":"HTTPValidationError","type":"object","properties":{"detail":{"title":"Detail","type":"array","items":{"$ref":"#/components/schemas/ValidationError"}}}},"ValidationError":{"title":"ValidationError","required":["loc","msg","type"],"type":"object","properties":{"loc":{"title":"Location","type":"array","items":{"anyOf":[{"type":"string"},{"type":"integer"}]}},"msg":{"title":"Message","type":"string"},"type":{"title":"Error Type","type":"string"}}},"VehicleDID":{"title":"VehicleDID","required":["diag_name","frame_id","start_bit"],"type":"object","properties":{"diag_name":{"title":"Diag Name","type":"string"},"frame_id":{"title":"Frame Id","type":"integer"},"start_bit":{"title":"Start Bit","type":"integer"},"hex_data":{"title":"Hex Data","type":"string"}}},"VehicleLogsObject":{"title":"VehicleLogsObject","required":["time","raw_data","diag_name","frame_id","start_bit","vin","decoded_data"],"type":"object","properties":{"time":{"title":"Time","type":"string","format":"date-time","example":"time"},"raw_data":{"title":"Raw Data","type":"string","example":"raw_data"},"diag_name":{"title":"Diag Name","type":"string","example":"diag_name"},"frame_id":{"title":"Frame Id","type":"integer","example":"frame_id"},"start_bit":{"title":"Start Bit","type":"integer","example":"start_bit"},"vin":{"title":"Vin","type":"string","example":"vin"},"decoded_data":{"title":"Decoded Data","type":"object","example":"decoded_data"}}},"VehicleModelDIDs":{"title":"VehicleModelDIDs","required":["model_id","model_name","vehicle_id","chip_id","dids_list"],"type":"object","properties":{"model_id":{"title":"Model Id","type":"string"},"model_name":{"title":"Model Name","type":"string"},"vehicle_id":{"title":"Vehicle Id","type":"string"},"chip_id":{"title":"Chip Id","type":"string"},"dids_list":{"title":"Dids List","type":"array","items":{"$ref":"#/components/schemas/VehicleDID"}}}}}}}